requirements.txt
torch==2.1.0
torchvision==0.16.0
fastapi==0.110.0
uvicorn==0.29.0
mlflow==2.9.2
pillow==10.2.0
pytest==7.4.3
